requirements.txt
marker-pdf
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
pydantic>=1.10.7
google-cloud-storage>=2.0.0
runpod